+++ libftprintf.h
@@ -6,7 +6,7 @@
 /*   By: fporciel <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/06 09:48:08 by fporciel          #+#    #+#             */
-/*   Updated: 2023/03/06 15:05:00 by fporciel         ###   ########.fr       */
+/*   Updated: 2023/03/07 09:26:51 by fporciel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ int		ft_is_format_spec(const char *format, int i);
 int		ft_conversion(const char *format, int *i, va_list ap);
 int		ft_putchar(int ap);
 int		ft_putstr(char *ap);
-int		ft_putaddress(uintptr_t ap);
+int		ft_putaddress(int ap);
 int		ft_putnbr(int ap);
 int		ft_putunbr(unsigned int ap);
 int		ft_puthexnbr(unsigned int ap, char formati);
